requirements_frontend.txt
streamlit>=1.41.1
pandas>=2.2.0
google-cloud-run>=0.10.14
pydantic>=2.10.5
